Add sqlite3 integration
Implemented working with databases, more specifically
sqlite3 databases into the codebase. I added an abstract class
essentially to shape how all database utility objects will look.
This class is provided with essentials such as the constructor for
creating the database folder and database file if they do not exist;
creating the connection to the database and closing the connection to the
database.

Due to the addition of this class, the BotUtils class was born and this class
will handle all the information relating to the general information for the bot
(guilds, prefixes, etc...).

There has been an introduction of a new enum called 'database'. This enum contains
very nested and detailed information for each database can was proven useful when
defining the BotUtils class.

All of these tools came together to allow for guilds to be automatically added to and
removed from the database when the bot has joined and left a guild respectively.