Finish notes command implementation
Getting this finished was a bit of a challenge for some reason.
Might have to do with the fact that I was primarily working on
this at 3 in the morning with little sleep but alas, it seemed
to magically fixed itself when I woke up and try again! JavaScript
is such a weird little spaghetti language.

Had to make the returns for get queries return a promise instead of
trying to run everything in sync since that caused some flow of code
issues. I'm still trying to master asynchronization in general but Js
has made it *that* much more confusing.

For some reason, the client was removed from the parameters of the run
function in the messageCreate event file. Re-added it and some bugs have
been squashed.