Add colored roles feature
This new colored roles features will allow server owners to add very
cool colored roles to their server. It was coded in a way to add
much customizability. In order to add this customizability everything
had to be placed in a JSON file.

This JSON file had to be created through the use of a generic class
called JSONConfig. In the concstructor of this class file the file name,
json file directory and path of the JSON config file is allocated.
In this class there are also default methods to make the JSON file if
it doesn't already exist along with the JSON directory. If the file
already exists this method would throw a custom error (FileHandlingError)
stating that the file already exists. This error would then be ignored
in the constructor of the subclasses once the method is called.
Going forward, each subclass of the JSONConfig class will call the super
method and then call it's own method (JSONConfig#initConfig) to initialize
everything that will be needed by the file.

Back to the command, there are three subcommands so far (add, remove, init)
and there may be more in the future.

There were also a few more utils added to the GeneralUtils class to aid with
the implementation of this new feature. GeneralUtils#isHex, GeneralUtils#isEmoji,
GeneralUtils#emojiPrefix are the new utils that were added.